savegame: move initial version to TRX
