+++ src/libtrx/game/savegame/common.c
@@ -1,7 +1,18 @@
 #include "game/savegame.h"
 #include "log.h"
 
+static SAVEGAME_VERSION m_InitialVersion = VERSION_LEGACY;
 static int32_t m_BoundSlot = -1;
+
+SAVEGAME_VERSION Savegame_GetInitialVersion(void)
+{
+    return m_InitialVersion;
+}
+
+void Savegame_SetInitialVersion(const SAVEGAME_VERSION version)
+{
+    m_InitialVersion = version;
+}
 
 void Savegame_BindSlot(const int32_t slot_num)
 {
